rename: feed -> rss
feedはrssのリンクを踏んで得られるもの
