+++ src/core/mod.rs
@@ -1,3 +1,3 @@
 pub mod article;
-pub mod feed;
 pub mod link;
+pub mod rss;
